recompute from pilot position

+++ M3/app/RaceSimulate/WayPoints.h
@@ -18,26 +18,26 @@ class XcPoint
 {
 public:
     XcPoint();
-	XcPoint(const char *name, float lat, float lon, float alt, const char *desc = nullptr);
+	XcPoint(const char *name, double lat, double lon, double alt, const char *desc = nullptr);
 
 	const char *getName() { return this->name.c_str(); }
 	const char *getDescription() { return this->description.c_str(); }
-	float getLatitude() { return this->lat; }
-	float getLongitude() { return this->lon; }
-	float getAltitude() { return this->altitude; }
+	double getLatitude() { return this->lat; }
+	double getLongitude() { return this->lon; }
+	double getAltitude() { return this->altitude; }
 
 	void setName(const char *name);
 	void setDescription(const char *desc);
-	void setLatitude(float lat);
-	void setLongitude(float lon);
-	void setAltitude(float alt);
+	void setLatitude(double lat);
+	void setLongitude(double lon);
+	void setAltitude(double alt);
 
 protected:
     String name;
     String description;
-    float lat;
-    float lon;
-    float altitude;
+    double lat;
+    double lon;
+    double altitude;
 };
